6889-enable-insurance-tab-even-when-program-module-off
